Implemented Factory Smelting and also updated a lot of old factory code

+++ src/main/java/com/skelril/aurora/items/CustomItemsComponent.java
@@ -394,7 +394,7 @@ public class CustomItemsComponent extends BukkitComponent implements Listener {
                 String modifierString = map.get(ChatColor.RED + "Damage Modifier");
                 if (modifierString != null) {
                     try {
-                        modifier = Integer.parseInt(modifierString);
+                        modifier = Double.parseDouble(modifierString);
                     } catch (NumberFormatException ignored) {
                     }
                 }
